Adds no_terminal_background option
This adds g:afterglow_no_terminal_background option, defaults to 0.
If set to 1, then the background terminal color won't be changed by the colorscheme.

It checks if we're not running gvim or neovim-gtk. Maybe there are some other gui clients to check.

This is useful, for example, for tmux users that change the background for the active pane.